fixed critical bug for wallet management

+++ src/database/repository/wallet.repository.ts
@@ -9,8 +9,8 @@ import { CreateWalletDto } from '../dto/wallet.dto';
 export class WalletRepository {
   constructor(@InjectModel(Wallet.name) private walletModel: Model<Wallet>) {}
 
-  async getById(walletId: string | Types.ObjectId): Promise<Wallet> {
-    const wallet = await this.walletModel.findById(walletId).exec();
+  async getById(walletId: string | Types.ObjectId, userId: number): Promise<Wallet> {
+    const wallet = await this.walletModel.findOne({ _id: walletId, userId }).exec();
     if (!wallet) throw new NotFoundException(`Not found wallet by address ${walletId.toString()}`);
     return wallet;
   }
@@ -19,14 +19,14 @@ export class WalletRepository {
     return this.walletModel.find({ userId }).exec();
   }
 
-  async getByAddress(address: string): Promise<Wallet> {
-    const wallet = await this.walletModel.findOne({ address }).exec();
+  async getByAddress(address: string, userId: number): Promise<Wallet> {
+    const wallet = await this.walletModel.findOne({ address, userId }).exec();
     if (!wallet) throw new NotFoundException(`Not found wallet by address ${address}`);
     return wallet;
   }
 
-  async getByAddresses(addresses: string[]): Promise<Wallet[]> {
-    return this.walletModel.find({ address: { $in: addresses } }).exec();
+  async getByAddresses(addresses: string[], userId: number): Promise<Wallet[]> {
+    return this.walletModel.find({ address: { $in: addresses }, userId }).exec();
   }
 
   async getOrCreateWallet(createWalletDto: CreateWalletDto): Promise<Wallet> {
@@ -39,10 +39,10 @@ export class WalletRepository {
     return wallet;
   }
 
-  async setMainWallet(walletId: string | Types.ObjectId) {
-    await this.walletModel.updateMany({ _id: { $ne: walletId } }, { $unset: { isMain: '' } });
-    await this.walletModel.updateOne({ _id: walletId }, { $set: { isMain: true } }, { runValidators: true });
-    return await this.getById(walletId);
+  async setMainWallet(walletId: string | Types.ObjectId, userId: number) {
+    await this.walletModel.updateMany({ _id: { $ne: walletId }, userId }, { $unset: { isMain: '' } });
+    await this.walletModel.updateOne({ _id: walletId, userId }, { $set: { isMain: true } }, { runValidators: true });
+    return await this.getById(walletId, userId);
   }
 
   async getMainWalletPrivateKeyForUser(userId: number) {
@@ -51,7 +51,7 @@ export class WalletRepository {
     return wallet.privateKey;
   }
 
-  async deleteByAddress(address: string) {
-    await this.walletModel.deleteOne({ address }).exec();
+  async deleteByAddress(address: string, userId: number) {
+    await this.walletModel.deleteOne({ address, userId }).exec();
   }
 }
